refactor+fix: revision of job complete state management (#120)
Sending the finish job message to agents should only be executed when job is
in running stae. The existing code sends the message if server is done.
It's very likely that when server is done, job is in running state.
To ensure correctness, however, sending the finish job message is moved into
cond_completing() method in running state class.

Across several functions, the code for sending command to agent is repeated.
This redundancy is removed.

Finally, code is refactored to improve maintainability.