feat: more templates for tests (#183)
Added templates for job status polling, to make requests and check if the job is in a certain state. Added cleanup template, that removes started processes after a test is executed.
Added more examples for running a simple job and for running, updating, completing a job. Also run_tests method was updated in order to include job_status and cleanup_processes templates. Test types were updated for the same reason. Also test types were added since we need to know what template to use when the test config is received.

Co-authored-by: ragaia@cisco.com <ragaia@cisco.com>