fix: state transition from updating to running (#196)
In case of a job update, if an agent has new workers to deploy but there
is no worker to update, the agent sends to a controller RUNNING status
instead of UPDATED status. In that case, the controller ignores the
RUNNING status and fails to correctly transition its state back to
RUNNING. This makes the job's state stays in the UPDATING state.
To handle the issue, cond_running method is implemented in the UPDATING
state.