fix: job stuck in completing (#216)
From time to time, job state was stuck in completing due to broken pipe error. Due to a timing issue, server worker might finish it's work and exit the process, before agent send it's command. Therefore, when the inference serving is done, the server worker will just wait for commands from the agent.

Co-authored-by: ragaia@cisco.com <ragaia@cisco.com>