misc: improvements on integration tests (#222)
Added some improvements on the integration tests. Updated local and remote logs folder location, moved the local logs removal for when a new test is started, also added an info message with the location of these logs, reduced the verbosity of the ssh commands, improved ansible play names, added cleanup of the remote 'tail' processes and some cleanup in the tests_dype.py file.

Co-authored-by: ragaia@cisco.com <ragaia@cisco.com>