fix: requests stuck in recovery (#266)
Whenever there's a failed worker, server will continue to send requests to the broken pipeline / worker, causing the logic that calculates the max inflight requests and number of inflight requests to stop server sending further requests, blocking the job execution after a successful recovery. This happens because the server is not aware that some requests were lost, to recalibrate the logic that sets the tx allow event. For that reason, whenever there's a new config received by the worker, we reset the number of inflight requests to 0 and unblock the tx_allow_evt event.