fix: state transition on worker failure (#272)
A worker can fail while stopping or completing a job. In these two cases, we don't need to do anything but implement cond_recovery in the Stopping and Completing states. Also, while testing, is a scenario when running controller with llama config and starting a resnet example, one worker might hang and not terminate itself, resulting in ConnectionResetError while waiting for messages from multiprocessing pipe. Added code to properly handle that error.