fix: grpc error propagation (#280)
Whenever an error is produced in the controller, it will be propagated via gRPC connection to all agents. For the moment, we suppress these error at the controller level and log them. Also, since agents will have a streaming connection via gRPC, we added error handler on that part. For this part, we check the status code for controller unavailability. Also, since agent has coroutines that continuously send data to the controller, when the controller dies, errors will be raised since there's nothing at the other end of connection.