feat: suppress noisy congestion signals (#291)
In order to not autoscale too often, we need to implement a mechanism to limit scale out when congestion is identified. For that, we need to identify 3 consecutive congestion signals before deciding to scale out. Congestion count is reset whenever there's no congestion or when the scale out is done.