fix: planner pipeline cleanup (#321)
When using autoscaler with recovery, when doing scale out and a worker failure happens that can't be recovered, we remove the affected pipeline and update the job. Since planner keeps track of active pipelines, we need to update that data structure as well, to avoid scale_in issues. For that, whenever the config is updated by removing pipelines / workers, planner cleanup is done.