refactor: get agent resources on demand
Due to timing issues when agents are periodically sending their resources, we need to refactor the way controller handles agent resources. Instead of having a polling mechanism from agents to send their resources using a fixed interval, now controller asks for resources whenever it needs them. For that reason, code was refactored to get agent resources whenever a start or update command is received. We added a mechanism that sends an rpc command to all agents then it will wait for all agents to return their resources before continuing with start or update. Did cleanup on polling code on agent and resources classes.